+++ internal/provider/model/job_definition/filter/filter_mask.go
