+++ internal/provider/model/job_definition/input_option/parser/jsonpath_parser.go
