+++ internal/provider/model/job_definition/output_option/bigquery.go
